ready to done(beta version)

+++ src/main/java/com/telran/contacts/service/ContactDtoServiceImpl.java
@@ -1,7 +1,7 @@
 package com.telran.contacts.service;
 
 import com.telran.contacts.exception.DuplicateRuntimeException;
-import com.telran.contacts.exception.UserNotFoundRuntimeException;
+import com.telran.contacts.exception.NotFoundRuntimeException;
 import com.telran.contacts.model.dto.AddressDto;
 import com.telran.contacts.model.dto.ContactDto;
 import com.telran.contacts.model.entity.Address;
@@ -34,50 +34,62 @@ public class ContactDtoServiceImpl implements ContactDtoService {
     @Override
     @Transactional(readOnly = true)
     public List<ContactDto> getAll() {
-        return userRepository.findAll().stream().map(user -> getContact(user.getUserId())).collect(Collectors.toList());
+        return userRepository.findAll().stream()
+                .map(user -> getContact(user.getUserId()))
+                .collect(Collectors.toList());
     }
 
     @Override
     @Transactional
     public void addContact(ContactDto contactDTO) {
-        contactDTO.getPhoneNumbers().forEach(this::checkIfExistPhone);
-        checkIfExistEmail(contactDTO.getEmail());
-        User user = User.builder().fullName(contactDTO.getFullName())
-                .email(contactDTO.getEmail()).createdDate(LocalDateTime.now()).build();
+        List<String> numbersDto = contactDTO.getPhoneNumbers();
+        numbersDto.forEach(number -> {
+            PhoneNumber phoneNumber = phoneNumberRepository.findPhoneNumberByPhoneNumber(number);
+            if (phoneNumber != null){
+                throw new DuplicateRuntimeException(number + " already exist!");
+            }
+        });
+        String email = contactDTO.getEmail();
+        if (userRepository.findUserByEmail(email) != null) {
+            throw new DuplicateRuntimeException(email + " already exist");
+        }
+        User user = User.builder()
+                .fullName(contactDTO.getFullName())
+                .email(contactDTO.getEmail())
+                .createdDate(LocalDateTime.now()).build();
         userRepository.save(user);
         List<String> numbers = contactDTO.getPhoneNumbers();
-        List<PhoneNumber> phoneNumbers = numbers.stream().map(number -> PhoneNumber.builder()
-                .createdDate(LocalDateTime.now()).phoneNumber(number).user(user).build()).collect(Collectors.toList());
+        List<PhoneNumber> phoneNumbers = numbers.stream()
+                .map(number -> PhoneNumber
+                        .builder()
+                        .createdDate(LocalDateTime.now())
+                        .phoneNumber(number)
+                        .user(user).build())
+                .collect(Collectors.toList());
         phoneNumberRepository.saveAll(phoneNumbers);
         List<Address> addresses = convDtoToAddress(contactDTO.getAddresses(), user);
         addressRepository.saveAll(addresses);
     }
 
-    private void checkIfExistPhone(String number) {
-        PhoneNumber phoneNumber = phoneNumberRepository.findPhoneNumberByPhoneNumber(number);
-        if (phoneNumber != null) {
-            throw new DuplicateRuntimeException(number + " already exist");
-        }
-    }
-
-    private void checkIfExistEmail(String email) {
-        if (userRepository.findUserByEmail(email) != null) {
-            throw new DuplicateRuntimeException(email + " already exist");
-        }
-    }
-
     @Override                                  //TODO
     @Transactional
     public void updateContact(ContactDto contactDto) {
+        User userByEmail = userRepository.findUserByEmail(contactDto.getEmail());
         User user = userRepository.findById(contactDto.getUserId()).orElseThrow(() ->
-                new UserNotFoundRuntimeException("User is not found"));
+                new NotFoundRuntimeException("User is not found"));
+        if (contactDto.getUserId() != userByEmail.getUserId()){
+            throw new DuplicateRuntimeException(contactDto.getEmail() + " already exist!");
+        }
         user.setFullName(contactDto.getFullName());
         user.setEmail(contactDto.getEmail());
         updateAddresses(contactDto, user);
         phoneNumberRepository.deleteAllByUser(user);
-        List<PhoneNumber> numbers = contactDto.getPhoneNumbers().stream().map(number -> {
-            return PhoneNumber.builder().phoneNumber(number).user(user).createdDate(LocalDateTime.now()).build();
-        }).collect(Collectors.toList());
+        List<PhoneNumber> numbers = contactDto.getPhoneNumbers().stream()
+                .map(number -> PhoneNumber.builder()
+                          .phoneNumber(number)
+                          .user(user)
+                          .createdDate(LocalDateTime.now()).build()
+               ).collect(Collectors.toList());
         phoneNumberRepository.saveAll(numbers);
     }
 
@@ -86,9 +98,14 @@ public class ContactDtoServiceImpl implements ContactDtoService {
         addressDtoList.forEach(addressDto -> {
             Address address = addressRepository.findById(addressDto.getAddressId()).orElse(null);
             if (address == null) {
-                addressRepository.save(Address.builder().country(addressDto.getCountry()).city(addressDto.getCity())
-                        .street(addressDto.getStreet()).houseNumber(addressDto.getHouseNumber()).apartment(addressDto.getApartment())
-                        .createdDate(LocalDateTime.now()).user(user).build());
+                addressRepository.save(Address.builder()
+                        .country(addressDto.getCountry())
+                        .city(addressDto.getCity())
+                        .street(addressDto.getStreet())
+                        .houseNumber(addressDto.getHouseNumber())
+                        .apartment(addressDto.getApartment())
+                        .createdDate(LocalDateTime.now())
+                        .user(user).build());
             }
             address.setApartment(addressDto.getApartment());
             address.setHouseNumber(addressDto.getHouseNumber());
@@ -102,7 +119,7 @@ public class ContactDtoServiceImpl implements ContactDtoService {
     @Transactional
     public void deleteContact(Long userId) {
         User user = userRepository.findById(userId).orElseThrow(() ->
-                new UserNotFoundRuntimeException("User is not exist by id : " + userId));
+                new NotFoundRuntimeException("User is not exist by id : " + userId));
         addressRepository.deleteAllByUser(user);
         phoneNumberRepository.deleteAllByUser(user);
         userRepository.deleteById(userId);
@@ -113,7 +130,7 @@ public class ContactDtoServiceImpl implements ContactDtoService {
     public ContactDto searchByPhoneNumber(String phoneNumber) {
         PhoneNumber phone = phoneNumberRepository.findPhoneNumberByPhoneNumber(phoneNumber);
         if (phone == null) {
-            return new ContactDto();
+            throw new NotFoundRuntimeException("User is not exist by phoneNumber : " + phoneNumber);
         }
         return getContact(phone.getUser().getUserId());
     }
@@ -125,43 +142,65 @@ public class ContactDtoServiceImpl implements ContactDtoService {
         if (users == null) {
             return new ArrayList<>();
         }
-        return users.stream().map(user -> getContact(user.getUserId())).collect(Collectors.toList());
+        return users.stream()
+                .map(user -> getContact(user.getUserId()))
+                .collect(Collectors.toList());
     }
 
     @Override
     @Transactional(readOnly = true)
     public ContactDto getContact(Long id) {
         User user = userRepository.findById(id).orElseThrow(() ->
-                new UserNotFoundRuntimeException("User is not exist by id : " + id));
+                new NotFoundRuntimeException("User is not exist by id : " + id));
         List<String> numbersDto = getNumbers(user);
         List<Address> addresses = addressRepository.findAllByUser(user);
 
-        return ContactDto.builder().userId(user.getUserId()).fullName(user.getFullName()).email(user.getEmail())
-                .phoneNumbers(numbersDto).addresses(convAddressToDto(addresses)).build();
+        return ContactDto.builder()
+                .userId(user.getUserId())
+                .fullName(user.getFullName())
+                .email(user.getEmail())
+                .phoneNumbers(numbersDto)
+                .addresses(convAddressToDto(addresses)).build();
     }
 
     private List<String> getNumbers(User user) {
         List<PhoneNumber> numbers = phoneNumberRepository.findAllByUser(user);
-        return numbers.stream().map(phoneNumber -> phoneNumber.getPhoneNumber()).collect(Collectors.toList());
+        return numbers.stream()
+                .map(phoneNumber -> phoneNumber.getPhoneNumber())
+                .collect(Collectors.toList());
     }
 
     public List<AddressDto> convAddressToDto(List<Address> addresses) {
         if (addresses.isEmpty()) {
             return new ArrayList<>();
         }
-        return addresses.stream().map(address -> AddressDto.builder().addressId(address.getAddressId()).country(address.getCountry())
-                .city(address.getCity()).street(address.getStreet()).houseNumber(address.getHouseNumber())
-                .apartment(address.getApartment()).build()).collect(Collectors.toList());
+        return addresses.stream()
+                .map(address -> AddressDto.builder()
+                        .addressId(address.getAddressId())
+                        .country(address.getCountry())
+                        .city(address.getCity())
+                        .street(address.getStreet())
+                        .houseNumber(address.getHouseNumber())
+                        .apartment(address.getApartment()).build())
+                .collect(Collectors.toList());
     }
 
     public List<Address> convDtoToAddress(List<AddressDto> addressDtoList, User user) {
         if (addressDtoList.isEmpty()) {
             return new ArrayList<>();
         }
-        return addressDtoList.stream().map(addressDto -> Address.builder().addressId(addressDto.getAddressId())
-                .user(user).country(addressDto.getCountry()).city(addressDto.getCity()).street(addressDto.getStreet())
-                .houseNumber(addressDto.getHouseNumber()).apartment(addressDto.getApartment()).createdDate(LocalDateTime.now())
-                .build()).collect(Collectors.toList());
+        return addressDtoList.stream()
+                .map(addressDto -> Address
+                        .builder()
+                        .addressId(addressDto.getAddressId())
+                        .user(user)
+                        .country(addressDto.getCountry())
+                        .city(addressDto.getCity())
+                        .street(addressDto.getStreet())
+                        .houseNumber(addressDto.getHouseNumber())
+                        .apartment(addressDto.getApartment())
+                        .createdDate(LocalDateTime.now()).build())
+                .collect(Collectors.toList());
     }
 }
